+++ esp-hal/src/soc/psram_common.rs
@@ -25,7 +25,7 @@ impl PsramSize {
 }
 
 /// Returns the address and size of the available in external memory.
-#[cfg(any(feature = "quad-psram", feature = "octal-psram"))]
+#[cfg(feature = "psram")]
 pub fn psram_raw_parts(_psram: &crate::peripherals::PSRAM) -> (*mut u8, usize) {
     let range = crate::soc::psram_range();
     (range.start as *mut u8, range.end - range.start)